// src/app/model/interview.ts
export class Interview {
    id: string | undefined;
    dateOfInterview: string | undefined;
    comment: string | undefined;
    interest: string | undefined;
    evaluation: string | undefined;
    selectEmployment: string | undefined;
    dateOfEmployment: string | undefined;
    candidateId: number | undefined;
}